-Moved the work done in App.tsx to a seperate component named Doggo.tsx
