+++ package.json
@@ -10,10 +10,12 @@
     "@types/node": "^12.20.33",
     "@types/react": "^17.0.30",
     "@types/react-dom": "^17.0.10",
+    "@types/react-router-dom": "^5.3.2",
     "axios": "^0.23.0",
     "node-fetch": "^3.0.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.1",
     "typescript": "^4.4.4",
     "web-vitals": "^1.1.2"
