+++ src/App.tsx
@@ -1,39 +1,19 @@
 import React, {useEffect} from 'react';
 import './App.css';
 import {DogService} from "./Services/Dog.Service";
-import {Dog} from "./Types/Dog";
+import {Dog} from "./Models/Dog";
+import Doggo from "./Components/Pages/Doggo"
 
-const dogService: DogService = new DogService("https://random.dog/woof.json");
+
 
 function App() {
-    let [dogs, setdogs] = React.useState<Dog>({fileSize: "", url: ""});
-    useEffect(() => {
-        getDog()
-    }, []);
 
-    const getDog = () => {
-        dogService.GET<Dog>().then(response => setdogs(response))
-        console.log(dogs)
-    }
 
     return (
+
+
         <div className="App">
-            <header className="App-header">
-                {
-                    dogs.url !== ""
-                    ? <img src={dogs.url} height={"250px"} width={"250px"} className="App-logo" alt="logo"/>
-                    : <img src="https://i.pinimg.com/originals/36/b1/6c/36b16cc68bbbca07f1c37183112de693.gif" className="App-logo" alt="logo"/>
-                }
-
-                {/*<List items={[*/}
-                {/*    {key: "1", name: "SASelenium"},*/}
-                {/*    {key: "2", name: "QA-Api-Testing"}*/}
-                {/*]} renderer={(x) => <div>{x.name}</div>}/>*/}
-
-                <button onClick={() => {getDog()}}>
-                    Click
-                </button>
-            </header>
+            <Doggo/>
         </div>
     );
 }
